Resolved a TODO
It was the case in the past that scaled and unscaled data channels were swapped, hence this fix.

A conversation with Tim.

Tim confirmed just now that for the new set of data we've got this is no longer the case, so we are removing this workaround.
Runs off the old data files should not be affected, because we always normalise the data.